+++ src/auth.rs
@@ -5,7 +5,7 @@ use thiserror::Error;
 use tonic::{codegen::http::uri::InvalidUri, transport::Channel, Streaming};
 
 use crate::{
-    response::MomentoError,
+    requests::{MomentoError, MomentoErrorCode},
     utils::{connect_channel_lazily, ChannelConnectError},
 };
 
@@ -67,29 +67,41 @@ pub enum AuthError {
 impl From<AuthError> for MomentoError {
     fn from(err: AuthError) -> Self {
         match err {
-            AuthError::LoginAborted => MomentoError::Cancelled {
-                description: "aborted login".into(),
-                source: err.into(),
+            AuthError::LoginAborted => MomentoError {
+                message: "aborted login".into(),
+                error_code: MomentoErrorCode::CancelledError,
+                inner_error: Some(err.into()),
+                details: None,
             },
-            AuthError::LoginFailed(_) => MomentoError::PermissionDenied {
-                description: "login failed".into(),
-                source: err.into(),
+            AuthError::LoginFailed(_) => MomentoError {
+                message: "login failed".into(),
+                error_code: MomentoErrorCode::PermissionError,
+                inner_error: Some(err.into()),
+                details: None,
             },
-            AuthError::BadUri(_) => MomentoError::BadRequest {
-                description: "bad uri".into(),
-                source: Some(err.into()),
+            AuthError::BadUri(_) => MomentoError {
+                message: "bad uri".into(),
+                error_code: MomentoErrorCode::BadRequestError,
+                inner_error: Some(err.into()),
+                details: None,
             },
-            AuthError::Connection(_) => MomentoError::InternalServerError {
-                description: "connection failed".into(),
-                source: err.into(),
+            AuthError::Connection(_) => MomentoError {
+                message: "connection failed".into(),
+                error_code: MomentoErrorCode::InternalServerError,
+                inner_error: Some(err.into()),
+                details: None,
             },
-            AuthError::ServerError(_) => MomentoError::InternalServerError {
-                description: "server error".into(),
-                source: err.into(),
+            AuthError::ServerError(_) => MomentoError {
+                message: "server error".into(),
+                error_code: MomentoErrorCode::InternalServerError,
+                inner_error: Some(err.into()),
+                details: None,
             },
-            AuthError::ActionError(_) => MomentoError::ClientSdkError {
-                description: "login action failed".into(),
-                source: err.into(),
+            AuthError::ActionError(_) => MomentoError {
+                message: "login action failed".into(),
+                error_code: MomentoErrorCode::UnknownError,
+                inner_error: Some(err.into()),
+                details: None,
             },
         }
     }