+++ src/requests/mod.rs
@@ -1,1 +1,4 @@
 pub mod cache;
+pub mod errors;
+
+pub use self::errors::*;